[refactor] added l10n.yaml file
