+++ l10n.yaml
@@ -1,0 +1,6 @@
+arb-dir: lib/l10n
+template-arb-file: app_pt.arb
+output-localization-file: (NOME DO ARQUIVO - EX: core_strings.dart)
+output-class: (NOME DA CLASSE - EX: CoreStrings)
+output-dir: lib/l10n
+synthetic-package: false
