+++ lib/crypto_wallet/core/routes/app_routes.dart
@@ -1,4 +1,4 @@
-import 'package:cryptowallet/app/presenter/ui/wallet/wallet_page.dart';
+import '../../presenter/ui/wallet/wallet_page.dart';
 
 var appRoutes = {
   WalletPage.walletRoute: (context) => const WalletPage(),
