+++ lib/crypto_wallet/core/utils/images.dart
@@ -1,0 +1,3 @@
+const String bitcoinIcon = 'assets/icons/bitcoinIcon.png';
+const String ethereumIcon = 'assets/icons/ethereumIcon.png';
+const String litecoinIcon = 'assets/icons/litecoinIcon.png';
