+++ lib/crypto_wallet/core/utils/text_styles.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+const cryptoSubtitleStyle = TextStyle(
+  fontFamily: 'Montserrat',
+  fontStyle: FontStyle.normal,
+  fontWeight: FontWeight.w700,
+  fontSize: 20,
+  color: Color.fromARGB(153, 47, 46, 50),
+);
+
+const currencyTextStyle = TextStyle(
+  fontFamily: 'Montserrat',
+  fontStyle: FontStyle.normal,
+  fontWeight: FontWeight.w700,
+  fontSize: 32,
+  color: Color.fromRGBO(47, 46, 50, 1),
+);
+
+const hideTextStyle = TextStyle(
+  fontFamily: 'Montserrat',
+  fontStyle: FontStyle.normal,
+  fontWeight: FontWeight.w700,
+  fontSize: 32,
+  color: Color.fromRGBO(47, 46, 50, 1),
+);
+
+const titleTextStyle = TextStyle(
+  fontFamily: 'Montserrat',
+  fontStyle: FontStyle.normal,
+  fontWeight: FontWeight.w700,
+  fontSize: 32,
+  color: Color.fromRGBO(224, 43, 87, 1),
+);
+
+const String hideCryptoAmount = "*****";
