+++ lib/crypto_wallet/presenter/ui/shared_widgets/crypto_bottom_navbar.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class CryptoBottomNavbar extends StatelessWidget {
+  const CryptoBottomNavbar({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return BottomNavigationBar(
+      currentIndex: 0,
+      elevation: 10,
+      selectedItemColor: const Color.fromRGBO(224, 43, 87, 1),
+      items: const [
+        BottomNavigationBarItem(
+          icon: ImageIcon(
+            AssetImage('assets/icons/warrenIcon.png'),
+          ),
+          label: 'Porfólio',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.currency_exchange),
+          label: 'Movimentações',
+        ),
+      ],
+    );
+  }
+}
