+++ lib/crypto_wallet/presenter/ui/shared_widgets/hide_show_button.dart
