+++ pubspec.yaml
@@ -37,6 +37,8 @@ dependencies:
   hooks_riverpod: ^1.0.4
   dio: ^4.0.6
   intl: ^0.17.0
+  flutter_localizations:
+    sdk: flutter
 
 dev_dependencies:
   flutter_test:
@@ -58,6 +60,7 @@ flutter:
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
+  generate: true
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
